# openapi.yaml
openapi: 3.0.3
info:
  title: AffiDeck API
  description: |
    # AffiDeck Backend API
    
    A comprehensive API for an affiliate marketing platform with marketplace, capital lending, communities, and referral capabilities.
    
    ## Authentication
    
    Most endpoints require token-based authentication using Bearer tokens. After login or registration, use the returned `access_token` in the `Authorization: Bearer {token}` header.
    
    Tokens expire after 15 minutes; use the `refresh_token` to obtain new tokens.
    
    ## Pagination
    
    List endpoints support `?page=&per_page=` query parameters. Responses include pagination metadata.
    
    ## Error Responses
    
    All errors return a JSON object with:
    - `message`: Human-readable error message
    - `errors` (optional): Field-specific validation errors
  version: 1.0.0
  contact:
    name: AffiDeck API Support
servers:
  - url: http://localhost/api
    description: Local development
  - url: https://api.affideck.com
    description: Production
security:
  - BearerAuth: []
paths:
...
  /offers:
    get:
      summary: List offers (public with optional auth enhancements)
      tags: [Offers]
      security: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - name: category
          in: query
          schema:
            type: string
        - name: geo
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, popular, highest_payout]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated offers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Offer'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      summary: Create a new offer (advertiser/partner only)
      tags: [Offers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, payout]
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum: [cpa, cpc, cpv]
                payout:
                  type: number
                  format: double
                tags:
                  type: array
                  items:
                    type: string
                categories:
                  type: array
                  items:
                    type: string
                geo:
                  type: array
                  items:
                    type: string
                creatives:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Offer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Offer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /communities:
    get:
      summary: List communities
      tags: [Communities]
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated communities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Community'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      summary: Create new community
      tags: [Communities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                private:
                  type: boolean
      responses:
        '201':
          description: Community created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Community'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /referrals:
    get:
      summary: List referrals
      tags: [Referrals]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Paginated referrals
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Referral'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /settings:
    get:
      summary: Get user settings
      tags: [Settings]
      responses:
        '200':
          description: User settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Offer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [cpa, cpc, cpv]
        payout:
          type: number
        owner_id:
          type: integer
        status:
          type: string
          enum: [draft, published, archived]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Community:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        owner_id:
          type: integer
        private:
          type: boolean
        member_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    Referral:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        referrer_id:
          type: integer
        created_at:
          type: string
          format: date-time
...
    UserSettings:
      type: object
      properties:
        theme:
          type: string
        language:
          type: string
        notifications_enabled:
          type: boolean
        two_factor_enabled:
          type: boolean
...
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        last_page:
          type: integer
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 15
        maximum: 100
...
  responses:
    Unauthorized:
      description: Unauthorized (missing or invalid token)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthenticated.
    Forbidden:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: This action is unauthorized.
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...